requirements.txt
Flask==1.0.2
SQLAlchemy==1.3.2
psycopg2==2.8
flask_sqlalchemy==2.3.1
flask_migrate==2.4.0
marshmallow==2.19.2
flask-marshmallow==0.10.0
marshmallow-sqlalchemy==0.16.2
flask-bcrypt==0.7.1
kazoo==2.6.0
requests==2.21.0
cryptography==2.6.1
pyjwt==1.7.1
